+++ tests/phpunit/tests/ProjectLocator.php
@@ -7,14 +7,13 @@
 
 namespace Required\Traduttore\Tests;
 
-use \GP_UnitTestCase;
 use \Required\Traduttore\Project;
 use \Required\Traduttore\ProjectLocator as Locator;
 
 /**
  * Test cases for \Required\Traduttore\ProjectLocator.
  */
-class ProjectLocator extends GP_UnitTestCase {
+class ProjectLocator extends TestCase {
 	/**
 	 * @var \GP_Project
 	 */
@@ -51,6 +50,38 @@ class ProjectLocator extends GP_UnitTestCase {
 				'source_url_template' => 'https://github.com/wearerequired/traduttore/blob/master/%file%#L%line%',
 			]
 		);
+	}
+
+	public function test_empty_string(): void {
+		$locator = new Locator( '' );
+
+		$this->assertNull( $locator->get_project() );
+	}
+
+	public function test_false(): void {
+		$locator = new Locator( false );
+
+		$this->assertNull( $locator->get_project() );
+	}
+
+	public function test_invalid_project_id(): void {
+		$locator = new Locator( 0 );
+
+		$this->assertNull( $locator->get_project() );
+	}
+
+	public function test_existing_project_instance(): void {
+		$project = new Project( $this->root );
+		$locator = new Locator( $project );
+
+		$this->assertSame( $project, $locator->get_project() );
+	}
+
+	public function test_existing_glotpress_project_instance(): void {
+		$project = $this->root;
+		$locator = new Locator( $project );
+
+		$this->assertEquals( $this->root->id, $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_glotpress_path(): void {