+++ tests/phpunit/tests/WebhookHandler/GitLab.php
@@ -7,17 +7,15 @@
 
 namespace Required\Traduttore\Tests\WebhookHandler;
 
-use \GP_UnitTestCase;
 use Required\Traduttore\Project;
 use Required\Traduttore\Repository;
-use WP_Error;
+use Required\Traduttore\Tests\TestCase;
 use \WP_REST_Request;
-use \WP_REST_Response;
 
 /**
  * Test cases for \Required\Traduttore\WebhookHandler\GitLab.
  */
-class GitLab extends GP_UnitTestCase {
+class GitLab extends TestCase {
 	/**
 	 * @var Project
 	 */
@@ -34,27 +32,6 @@ class GitLab extends GP_UnitTestCase {
 				]
 			)
 		);
-	}
-
-	/**
-	 * @see WP_Test_REST_TestCase
-	 *
-	 * @param mixed                     $code
-	 * @param WP_REST_Response|WP_Error $response
-	 * @param mixed                     $status
-	 */
-	protected function assertErrorResponse( $code, $response, $status = null ): void {
-		if ( $response instanceof WP_REST_Response ) {
-			$response = $response->as_error();
-		}
-
-		$this->assertInstanceOf( 'WP_Error', $response );
-		$this->assertEquals( $code, $response->get_error_code() );
-		if ( null !== $status ) {
-			$data = $response->get_error_data();
-			$this->assertArrayHasKey( 'status', $data );
-			$this->assertEquals( $status, $data['status'] );
-		}
 	}
 
 	public function test_missing_event_header(): void {
@@ -148,6 +125,40 @@ class GitLab extends GP_UnitTestCase {
 		);
 		$request->add_header( 'x-gitlab-event', 'Push Hook' );
 		$request->add_header( 'x-gitlab-token', 'traduttore-test' );
+		$response = rest_get_server()->dispatch( $request );
+
+		$this->assertEquals( 200, $response->get_status() );
+		$this->assertSame( [ 'result' => 'OK' ], $response->get_data() );
+		$this->assertSame( Repository::VCS_TYPE_GIT, $this->project->get_repository_vcs_type() );
+		$this->assertSame( Repository::TYPE_GITLAB, $this->project->get_repository_type() );
+		$this->assertSame( 'wearerequired/traduttore', $this->project->get_repository_name() );
+		$this->assertSame( 'https://gitlab.com/wearerequired/traduttore', $this->project->get_repository_url() );
+		$this->assertSame( 'git@gitlab.com:wearerequired/traduttore.git', $this->project->get_repository_ssh_url() );
+		$this->assertSame( 'https://gitlab.com/wearerequired/traduttore.git', $this->project->get_repository_https_url() );
+		$this->assertSame( 'public', $this->project->get_repository_visibility() );
+	}
+
+	public function test_valid_project_custom_webhook_secret(): void {
+		$secret = 'Sup3rS3cr3tPassw0rd';
+
+		$this->project->set_repository_webhook_secret( $secret );
+
+		$request = new WP_REST_Request( 'POST', '/traduttore/v1/incoming-webhook' );
+		$request->set_body_params(
+			[
+				'ref'     => 'refs/heads/master',
+				'project' => [
+					'default_branch'      => 'master',
+					'path_with_namespace' => 'wearerequired/traduttore',
+					'homepage'            => 'https://gitlab.com/wearerequired/traduttore',
+					'http_url'            => 'https://gitlab.com/wearerequired/traduttore.git',
+					'ssh_url'             => 'git@gitlab.com:wearerequired/traduttore.git',
+					'visibility_level'    => 0,
+				],
+			]
+		);
+		$request->add_header( 'x-gitlab-event', 'Push Hook' );
+		$request->add_header( 'x-gitlab-token', $secret );
 		$response = rest_get_server()->dispatch( $request );
 
 		$this->assertEquals( 200, $response->get_status() );