+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added new REST API route for incoming webhooks (`traduttore/v1/incoming-webhook`) which replaces the deprecated route (`github-webhook/v1/push-event`)
 * Changed all filters and actions to use `.` as the separator between the prefix and hook name instead of `_`
 * Improved scheduling of cron events to reduce number of unnecessary builds and updates
+* Added new `wp traduttore info` CLI command
 * Greatly improved documentation
 
 ## 2.0.3